+++ example/src/views/SinglePlayer.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Wrapper from '../components/Wrapper';
+import Player from '../components/Player';
+
+const SinglePlayer = () => {
+    return (
+        <Wrapper>
+            <Player />
+        </Wrapper>
+    );
+};
+
+export default SinglePlayer;
